@@ -1,20 +1,17 @@
 name: flutter_launcher_icons
 description: A package which simplifies the task of updating your Flutter app's launcher icon.
-version: 0.7.4
+version: 0.9.0
 maintainer: Mark O'Sullivan (@MarkOSullivan94)
 homepage: https://github.com/fluttercommunity/flutter_launcher_icons
-authors: 
-- Flutter Community <community@flutter.zone>
-- Franz Silva <franzsilva@gmail.com>
-- Mark O'Sullivan <mos182@hotmail.co.uk>
 
-dependencies: 
-  image: "^2.1.1"
-  args: "^1.5.0"
-  yaml: "^2.1.15"
+dependencies:
+  args: 2.0.0
+  image: ^3.0.0-nullsafety.0
+  path: ^1.8.0-nullsafety.3
+  yaml: ^3.0.0-nullsafety.0
 
 environment:
-  sdk: '>=2.2.0 <3.0.0'
+  sdk: '>=2.12.0-0 <3.0.0'
 
 dev_dependencies:
-  test: '^1.6.3'
+  test: ^1.16.0-nullsafety.16
