@@ -1,11 +1,13 @@
 import 'dart:io';
+import 'package:flutter_launcher_icons/utils.dart';
 import 'package:flutter_launcher_icons/xml_templates.dart' as xml_template;
 import 'package:image/image.dart';
 import 'package:flutter_launcher_icons/custom_exceptions.dart';
 import 'package:flutter_launcher_icons/constants.dart' as constants;
 
 class AndroidIconTemplate {
-  AndroidIconTemplate({this.size, this.directoryName});
+  AndroidIconTemplate({required this.size, required this.directoryName});
+
   final String directoryName;
   final int size;
 }
@@ -26,26 +28,33 @@ List<AndroidIconTemplate> androidIcons = <AndroidIconTemplate>[
   AndroidIconTemplate(directoryName: 'mipmap-xxxhdpi', size: 192),
 ];
 
-void createDefaultIcons(Map<String, dynamic> flutterLauncherIconsConfig) {
-  print('Creating default icons Android');
+void createDefaultIcons(
+    Map<String, dynamic> flutterLauncherIconsConfig, String? flavor) {
+  printStatus('Creating default icons Android');
   final String filePath = getAndroidIconPath(flutterLauncherIconsConfig);
-  final Image image = decodeImage(File(filePath).readAsBytesSync());
+  final Image? image = decodeImageFile(filePath);
+  if (image == null) {
+    return;
+  }
+  final File androidManifestFile = File(constants.androidManifestFile);
   if (isCustomAndroidFile(flutterLauncherIconsConfig)) {
-    print('Adding a new Android launcher icon');
+    printStatus('Adding a new Android launcher icon');
     final String iconName = getNewIconName(flutterLauncherIconsConfig);
     isAndroidIconNameCorrectFormat(iconName);
     final String iconPath = '$iconName.png';
     for (AndroidIconTemplate template in androidIcons) {
-      saveNewImages(template, image, iconPath);
+      saveNewImages(template, image, iconPath, flavor);
     }
-    overwriteAndroidManifestWithNewLauncherIcon(iconName);
+    overwriteAndroidManifestWithNewLauncherIcon(iconName, androidManifestFile);
   } else {
-    print('Overwriting the default Android launcher icon with a new icon');
+    printStatus(
+        'Overwriting the default Android launcher icon with a new icon');
     for (AndroidIconTemplate template in androidIcons) {
-      overwriteExistingIcons(template, image, constants.androidFileName);
+      overwriteExistingIcons(
+          template, image, constants.androidFileName, flavor);
     }
     overwriteAndroidManifestWithNewLauncherIcon(
-        constants.androidDefaultIconName);
+        constants.androidDefaultIconName, androidManifestFile);
   }
 }
 
@@ -59,28 +68,33 @@ bool isAndroidIconNameCorrectFormat(String iconName) {
   return true;
 }
 
-void createAdaptiveIcons(Map<String, dynamic> flutterLauncherIconsConfig) {
-  print('Creating adaptive icons Android');
+void createAdaptiveIcons(
+    Map<String, dynamic> flutterLauncherIconsConfig, String? flavor) {
+  printStatus('Creating adaptive icons Android');
 
   // Retrieve the necessary Flutter Launcher Icons configuration from the pubspec.yaml file
   final String backgroundConfig =
-  flutterLauncherIconsConfig['adaptive_icon_background'];
+      flutterLauncherIconsConfig['adaptive_icon_background'];
   final String foregroundImagePath =
-  flutterLauncherIconsConfig['adaptive_icon_foreground'];
-  final Image foregroundImage =
-  decodeImage(File(foregroundImagePath).readAsBytesSync());
+      flutterLauncherIconsConfig['adaptive_icon_foreground'];
+  final Image? foregroundImage = decodeImageFile(foregroundImagePath);
+  if (foregroundImage == null) {
+    return;
+  }
 
   // Create adaptive icon foreground images
   for (AndroidIconTemplate androidIcon in adaptiveForegroundIcons) {
-    overwriteExistingIcons(androidIcon, foregroundImage, constants.androidAdaptiveForegroundFileName);
+    overwriteExistingIcons(androidIcon, foregroundImage,
+        constants.androidAdaptiveForegroundFileName, flavor);
   }
 
   // Create adaptive icon background
   if (isAdaptiveIconConfigPngFile(backgroundConfig)) {
-    createAdaptiveBackgrounds(flutterLauncherIconsConfig, backgroundConfig);
+    createAdaptiveBackgrounds(
+        flutterLauncherIconsConfig, backgroundConfig, flavor);
   } else {
-    createAdaptiveIconMipmapXmlFile(flutterLauncherIconsConfig);
-    updateColorsXmlFile(backgroundConfig);
+    createAdaptiveIconMipmapXmlFile(flutterLauncherIconsConfig, flavor);
+    updateColorsXmlFile(backgroundConfig, flavor);
   }
 }
 
@@ -91,33 +105,35 @@ void createAdaptiveIcons(Map<String, dynamic> flutterLauncherIconsConfig) {
 ///
 /// If not, the colors.xml file is created and a color item for the adaptive icon
 /// background is included in the new colors.xml file.
-void updateColorsXmlFile(String backgroundConfig) {
-  final File colorsXml = File(constants.androidColorsFile);
+void updateColorsXmlFile(String backgroundConfig, String? flavor) {
+  final File colorsXml = File(constants.androidColorsFile(flavor));
   if (colorsXml.existsSync()) {
-    print('Updating colors.xml with color for adaptive icon background');
+    printStatus('Updating colors.xml with color for adaptive icon background');
     updateColorsFile(colorsXml, backgroundConfig);
   } else {
-    print('No colors.xml file found in your Android project');
-    print('Creating colors.xml file and adding it to your Android project');
-    createNewColorsFile(backgroundConfig);
+    printStatus('No colors.xml file found in your Android project');
+    printStatus(
+        'Creating colors.xml file and adding it to your Android project');
+    createNewColorsFile(backgroundConfig, flavor);
   }
 }
 
 /// Creates the xml file required for the adaptive launcher icon
 /// FILE LOCATED HERE: res/mipmap-anydpi/{icon-name-from-yaml-config}.xml
-void createAdaptiveIconMipmapXmlFile(Map<String, dynamic> flutterLauncherIconsConfig) {
+void createAdaptiveIconMipmapXmlFile(
+    Map<String, dynamic> flutterLauncherIconsConfig, String? flavor) {
   if (isCustomAndroidFile(flutterLauncherIconsConfig)) {
-    File(constants.androidAdaptiveXmlFolder +
-        getNewIconName(flutterLauncherIconsConfig) +
-        '.xml')
+    File(constants.androidAdaptiveXmlFolder(flavor) +
+            getNewIconName(flutterLauncherIconsConfig) +
+            '.xml')
         .create(recursive: true)
         .then((File adaptiveIcon) {
       adaptiveIcon.writeAsString(xml_template.icLauncherXml);
     });
   } else {
-    File(constants.androidAdaptiveXmlFolder +
-        constants.androidDefaultIconName +
-        '.xml')
+    File(constants.androidAdaptiveXmlFolder(flavor) +
+            constants.androidDefaultIconName +
+            '.xml')
         .create(recursive: true)
         .then((File adaptiveIcon) {
       adaptiveIcon.writeAsString(xml_template.icLauncherXml);
@@ -126,31 +142,35 @@ void createAdaptiveIconMipmapXmlFile(Map<String, dynamic> flutterLauncherIconsCo
 }
 
 /// creates adaptive background using png image
-void createAdaptiveBackgrounds(
-    Map<String, dynamic> yamlConfig, String adaptiveIconBackgroundImagePath) {
+void createAdaptiveBackgrounds(Map<String, dynamic> yamlConfig,
+    String adaptiveIconBackgroundImagePath, String? flavor) {
   final String filePath = adaptiveIconBackgroundImagePath;
-  final Image image = decodeImage(File(filePath).readAsBytesSync());
+  final Image? image = decodeImageFile(filePath);
+  if (image == null) {
+    return;
+  }
 
   // creates a png image (ic_adaptive_background.png) for the adaptive icon background in each of the locations
   // it is required
   for (AndroidIconTemplate androidIcon in adaptiveForegroundIcons) {
-    saveNewImages(androidIcon, image, constants.androidAdaptiveBackgroundFileName);
+    saveNewImages(androidIcon, image,
+        constants.androidAdaptiveBackgroundFileName, flavor);
   }
 
   // Creates the xml file required for the adaptive launcher icon
   // FILE LOCATED HERE:  res/mipmap-anydpi/{icon-name-from-yaml-config}.xml
   if (isCustomAndroidFile(yamlConfig)) {
-    File(constants.androidAdaptiveXmlFolder +
-        getNewIconName(yamlConfig) +
-        '.xml')
+    File(constants.androidAdaptiveXmlFolder(flavor) +
+            getNewIconName(yamlConfig) +
+            '.xml')
         .create(recursive: true)
         .then((File adaptiveIcon) {
       adaptiveIcon.writeAsString(xml_template.icLauncherDrawableBackgroundXml);
     });
   } else {
-    File(constants.androidAdaptiveXmlFolder +
-        constants.androidDefaultIconName +
-        '.xml')
+    File(constants.androidAdaptiveXmlFolder(flavor) +
+            constants.androidDefaultIconName +
+            '.xml')
         .create(recursive: true)
         .then((File adaptiveIcon) {
       adaptiveIcon.writeAsString(xml_template.icLauncherDrawableBackgroundXml);
@@ -159,8 +179,8 @@ void createAdaptiveBackgrounds(
 }
 
 /// Creates a colors.xml file if it was missing from android/app/src/main/res/values/colors.xml
-void createNewColorsFile(String backgroundColor) {
-  File(constants.androidColorsFile)
+void createNewColorsFile(String backgroundColor, String? flavor) {
+  File(constants.androidColorsFile(flavor))
       .create(recursive: true)
       .then((File colorsFile) {
     colorsFile.writeAsString(xml_template.colorsXml).then((File file) {
@@ -211,14 +231,17 @@ String getNewIconName(Map<String, dynamic> config) {
 /// Note: Do not change interpolation unless you end up with better results (see issue for result when using cubic
 /// interpolation)
 /// https://github.com/fluttercommunity/flutter_launcher_icons/issues/101#issuecomment-495528733
-void overwriteExistingIcons(AndroidIconTemplate template, Image image, String filename) {
-  Image newFile;
-  if (image.width > template.size) {
-    newFile = copyResize(image, width: template.size, height: -1, interpolation: Interpolation.average);
-  } else {
-    newFile = copyResize(image, width: template.size, height: -1, interpolation: Interpolation.linear);
-  }
-  File(constants.androidResFolder + template.directoryName + '/' + filename)
+void overwriteExistingIcons(
+  AndroidIconTemplate template,
+  Image image,
+  String filename,
+  String? flavor,
+) {
+  final Image newFile = createResizedImage(template.size, image);
+  File(constants.androidResFolder(flavor) +
+          template.directoryName +
+          '/' +
+          filename)
       .create(recursive: true)
       .then((File file) {
     file.writeAsBytesSync(encodePng(newFile));
@@ -228,14 +251,13 @@ void overwriteExistingIcons(AndroidIconTemplate template, Image image, String fi
 /// Saves new launcher icons to the project, keeping the old launcher icons.
 /// Note: Do not change interpolation unless you end up with better results
 /// https://github.com/fluttercommunity/flutter_launcher_icons/issues/101#issuecomment-495528733
-void saveNewImages(AndroidIconTemplate template, Image image, String iconFilePath) {
-  Image newFile;
-  if (image.width >= template.size) {
-    newFile = copyResize(image, width: template.size, height: template.size, interpolation: Interpolation.average);
-  } else {
-    newFile = copyResize(image, width: template.size, height: template.size, interpolation: Interpolation.linear);
-  }
-  File(constants.androidResFolder + template.directoryName + '/' + iconFilePath)
+void saveNewImages(AndroidIconTemplate template, Image image,
+    String iconFilePath, String? flavor) {
+  final Image newFile = createResizedImage(template.size, image);
+  File(constants.androidResFolder(flavor) +
+          template.directoryName +
+          '/' +
+          iconFilePath)
       .create(recursive: true)
       .then((File file) {
     file.writeAsBytesSync(encodePng(newFile));
@@ -246,11 +268,20 @@ void saveNewImages(AndroidIconTemplate template, Image image, String iconFilePat
 /// with the new icon name (only if it has changed)
 ///
 /// Note: default iconName = "ic_launcher"
-Future<void> overwriteAndroidManifestWithNewLauncherIcon(String iconName) async {
-  final File androidManifestFile = File(constants.androidManifestFile);
-  final List<String> lines = await androidManifestFile.readAsLines();
-  for (int x = 0; x < lines.length; x++) {
-    String line = lines[x];
+Future<void> overwriteAndroidManifestWithNewLauncherIcon(
+    String iconName, File androidManifestFile) async {
+  // we do not use `file.readAsLinesSync()` here because that always gets rid of the last empty newline
+  final List<String> oldManifestLines =
+      (await androidManifestFile.readAsString()).split('\n');
+  final List<String> transformedLines =
+      transformAndroidManifestWithNewLauncherIcon(oldManifestLines, iconName);
+  await androidManifestFile.writeAsString(transformedLines.join('\n'));
+}
+
+/// Updates only the line containing android:icon with the specified iconName
+List<String> transformAndroidManifestWithNewLauncherIcon(
+    List<String> oldManifestLines, String iconName) {
+  return oldManifestLines.map((String line) {
     if (line.contains('android:icon')) {
       // Using RegExp replace the value of android:icon to point to the new icon
       // anything but a quote of any length: [^"]*
@@ -259,14 +290,12 @@ Future<void> overwriteAndroidManifestWithNewLauncherIcon(String iconName) async 
       // repeat as often as wanted with no quote at start: [^"]*(\"[^"]*)*
       // escaping the slash to place in string: [^"]*(\\"[^"]*)*"
       // result: any string which does only include escaped quotes
-      line = line.replaceAll(RegExp(r'android:icon="[^"]*(\\"[^"]*)*"'),
+      return line.replaceAll(RegExp(r'android:icon="[^"]*(\\"[^"]*)*"'),
           'android:icon="@mipmap/$iconName"');
-      lines[x] = line;
-      // used to stop git showing a diff if the icon name hasn't changed
-      lines.add('');
+    } else {
+      return line;
     }
-  }
-  await androidManifestFile.writeAsString(lines.join('\n'));
+  }).toList();
 }
 
 /// Retrieves the minSdk value from the Android build.gradle file
@@ -277,7 +306,6 @@ int minSdk() {
     if (line.contains('minSdkVersion')) {
       // remove anything from the line that is not a digit
       final String minSdk = line.replaceAll(RegExp(r'[^\d]'), '');
-      print('Android minSdkVersion = $minSdk');
       return int.parse(minSdk);
     }
   }
