@@ -1,11 +1,13 @@
 import 'dart:convert';
 import 'dart:io';
+import 'package:flutter_launcher_icons/utils.dart';
 import 'package:image/image.dart';
 import 'package:flutter_launcher_icons/constants.dart';
 
 /// File to handle the creation of icons for iOS platform
 class IosIconTemplate {
-  IosIconTemplate({this.size, this.name});
+  IosIconTemplate({required this.size, required this.name});
+
   final String name;
   final int size;
 }
@@ -20,46 +22,60 @@ List<IosIconTemplate> iosIcons = <IosIconTemplate>[
   IosIconTemplate(name: '-40x40@1x', size: 40),
   IosIconTemplate(name: '-40x40@2x', size: 80),
   IosIconTemplate(name: '-40x40@3x', size: 120),
-  IosIconTemplate(name: '-50x50@1x', size: 50),
-  IosIconTemplate(name: '-50x50@2x', size: 100),
-  IosIconTemplate(name: '-57x57@1x', size: 57),
-  IosIconTemplate(name: '-57x57@2x', size: 114),
   IosIconTemplate(name: '-60x60@2x', size: 120),
   IosIconTemplate(name: '-60x60@3x', size: 180),
-  IosIconTemplate(name: '-72x72@1x', size: 72),
-  IosIconTemplate(name: '-72x72@2x', size: 144),
   IosIconTemplate(name: '-76x76@1x', size: 76),
   IosIconTemplate(name: '-76x76@2x', size: 152),
   IosIconTemplate(name: '-83.5x83.5@2x', size: 167),
   IosIconTemplate(name: '-1024x1024@1x', size: 1024),
 ];
 
-void createIcons(Map<String, dynamic> config) {
+void createIcons(Map<String, dynamic> config, String? flavor) {
   final String filePath = config['image_path_ios'] ?? config['image_path'];
-  final Image image = decodeImage(File(filePath).readAsBytesSync());
+  // decodeImageFile shows error message if null
+  // so can return here if image is null
+  final Image? image = decodeImage(File(filePath).readAsBytesSync());
+  if (image == null) {
+    return;
+  }
+  if (config['remove_alpha_ios'] is bool && config['remove_alpha_ios']) {
+    image.channels = Channels.rgb;
+  }
+  if (image.channels == Channels.rgba) {
+    print('\nWARNING: Icons with alpha channel are not allowed in the Apple App Store.\nSet "remove_alpha_ios: true" to remove it.\n');
+  }
   String iconName;
   final dynamic iosConfig = config['ios'];
-  // If the IOS configuration is a string then the user has specified a new icon to be created
-  // and for the old icon file to be kept
-  if (iosConfig is String) {
+  if (flavor != null) {
+    final String catalogName = 'AppIcon-$flavor';
+    printStatus('Building iOS launcher icon for $flavor');
+    for (IosIconTemplate template in iosIcons) {
+      saveNewIcons(template, image, catalogName);
+    }
+    iconName = iosDefaultIconName;
+    changeIosLauncherIcon(catalogName, flavor);
+    modifyContentsFile(catalogName);
+  } else if (iosConfig is String) {
+    // If the IOS configuration is a string then the user has specified a new icon to be created
+    // and for the old icon file to be kept
     final String newIconName = iosConfig;
-    print('Adding new iOS launcher icon');
+    printStatus('Adding new iOS launcher icon');
     for (IosIconTemplate template in iosIcons) {
       saveNewIcons(template, image, newIconName);
     }
     iconName = newIconName;
-    changeIosLauncherIcon(iconName);
+    changeIosLauncherIcon(iconName, flavor);
     modifyContentsFile(iconName);
   }
   // Otherwise the user wants the new icon to use the default icons name and
   // update config file to use it
   else {
-    print('Overwriting default iOS launcher icon with new icon');
+    printStatus('Overwriting default iOS launcher icon with new icon');
     for (IosIconTemplate template in iosIcons) {
       overwriteDefaultIcons(template, image);
     }
     iconName = iosDefaultIconName;
-    changeIosLauncherIcon('AppIcon');
+    changeIosLauncherIcon('AppIcon', flavor);
   }
 }
 
@@ -87,23 +103,47 @@ void saveNewIcons(IosIconTemplate template, Image image, String newIconName) {
 
 Image createResizedImage(IosIconTemplate template, Image image) {
   if (image.width >= template.size) {
-    return copyResize(image, width: template.size, height: template.size, interpolation: Interpolation.average);
+    return copyResize(image,
+        width: template.size,
+        height: template.size,
+        interpolation: Interpolation.average);
   } else {
-    return copyResize(image, width: template.size, height: template.size, interpolation: Interpolation.linear);
+    return copyResize(image,
+        width: template.size,
+        height: template.size,
+        interpolation: Interpolation.linear);
   }
 }
 
-Future<void> changeIosLauncherIcon(String iconName) async {
+Future<void> changeIosLauncherIcon(String iconName, String? flavor) async {
   final File iOSConfigFile = File(iosConfigFile);
   final List<String> lines = await iOSConfigFile.readAsLines();
+
+  bool onConfigurationSection = false;
+  String? currentConfig;
+
   for (int x = 0; x < lines.length; x++) {
-    String line = lines[x];
-    if (line.contains('ASSETCATALOG')) {
-      line = line.replaceAll(RegExp(r'=.*;'), '= $iconName;');
-      lines[x] = line;
-      lines[lines.length - 1] = '}\n';
+    final String line = lines[x];
+    if (line.contains('/* Begin XCBuildConfiguration section */')) {
+      onConfigurationSection = true;
+    }
+    if (line.contains('/* End XCBuildConfiguration section */')) {
+      onConfigurationSection = false;
+    }
+    if (onConfigurationSection) {
+      final match = RegExp('.*/\\* (.*)\.xcconfig \\*/;').firstMatch(line);
+      if (match != null) {
+        currentConfig = match.group(1);
+      }
+
+      if (currentConfig != null &&
+          (flavor == null || currentConfig.contains('-$flavor')) &&
+          line.contains('ASSETCATALOG')) {
+        lines[x] = line.replaceAll(RegExp('\=(.*);'), '= $iconName;');
+      }
     }
   }
+
   final String entireFile = lines.join('\n');
   await iOSConfigFile.writeAsString(entireFile);
 }
@@ -113,7 +153,8 @@ void modifyContentsFile(String newIconName) {
   final String newIconFolder =
       iosAssetFolder + newIconName + '.appiconset/Contents.json';
   File(newIconFolder).create(recursive: true).then((File contentsJsonFile) {
-    final String contentsFileContent = generateContentsFileAsString(newIconName);
+    final String contentsFileContent =
+        generateContentsFileAsString(newIconName);
     contentsJsonFile.writeAsString(contentsFileContent);
   });
 }
@@ -127,7 +168,13 @@ String generateContentsFileAsString(String newIconName) {
 }
 
 class ContentsImageObject {
-  ContentsImageObject({this.size, this.idiom, this.filename, this.scale});
+  ContentsImageObject({
+    required this.size,
+    required this.idiom,
+    required this.filename,
+    required this.scale,
+  });
+
   final String size;
   final String idiom;
   final String filename;
@@ -144,7 +191,8 @@ class ContentsImageObject {
 }
 
 class ContentsInfoObject {
-  ContentsInfoObject({this.version, this.author});
+  ContentsInfoObject({required this.version, required this.author});
+
   final int version;
   final String author;
 
@@ -269,7 +317,7 @@ List<Map<String, String>> createImageList(String fileNamePrefix) {
     ContentsImageObject(
             size: '1024x1024',
             idiom: 'ios-marketing',
-            filename: fileNamePrefix + '-1024x1024@1x.png',
+            filename: '$fileNamePrefix-1024x1024@1x.png',
             scale: '1x')
         .toJson()
   ];
